Modify TwoLayerNet: use Backpropagation

+++ src/main/scala/com/github/who_you_me/deeplearning/layer/SoftmaxWithLoss.scala
@@ -22,6 +22,6 @@ class SoftmaxWithLoss {
     val t = this.t.get
 
     val batchSize = t.rows
-    (y - t) / batchSize
+    (y - t) / batchSize.toDouble
   }
 }
